Shunting yard algorithm; rework of operator enum; doubles used now

+++ src/Expression.java
@@ -1,3 +1,3 @@
 public interface Expression {
-    int eval();
+    double eval();
 }
